requirements.txt
pyjwt~=2.10.1
Flask~=3.1.0
flask-cors~=5.0.1
sympy~=1.13.3
dotenv~=0.9.9
psycopg2
python-dotenv~=1.0.1
pycryptodome~=3.22.0
redis~=5.2.1